fix: mv mailchimp user tag updates to separate file
